+++ src/index.ts
@@ -184,7 +184,7 @@ class TheAlgorithm {
         //       so those Toot objects are completely built before we get here.
         newToots.forEach(toot => toot.setDependentProperties(userData, this.trendingData.links));
         this.feed = await this.cleanupFeed(newToots);
-        this.filters = initializeFiltersWithSummaryInfo(this.feed, await MastoApi.instance.getUserData());
+        this.filters = initializeFiltersWithSummaryInfo(this.feed, userData);
 
         // Potentially fetch more toots if we haven't reached the desired number
         this.maybeGetMoreToots(homeToots, numTimelineToots);  // Called asynchronously
@@ -246,6 +246,7 @@ class TheAlgorithm {
 
     // Remove invalid and duplicate toots
     private async cleanupFeed(toots: Toot[]): Promise<Toot[]> {
+        // TODO: could just retrieve mutedAccounts here...
         const mutedAccounts = (await MastoApi.instance.getUserData()).mutedAccounts || [];
         const cleanNewToots = toots.filter(toot => toot.isValidForFeed(mutedAccounts));
         const numRemoved = toots.length - cleanNewToots.length;