+++ src/scorer/feature/retoots_in_feed_scorer.ts
@@ -16,7 +16,7 @@ export default class RetootsInFeedScorer extends FeatureScorer {
         if (toot.reblog) {
             const reblog = toot.reblog;
             const nonAuthorRetoots = reblog.reblogsBy.filter((acct) => acct.webfingerURI != reblog.account.webfingerURI);
-            return nonAuthorRetoots.length;
+            return Math.pow(nonAuthorRetoots.length, 2);
         } else {
             return 0;
         }