+++ src/scorer/feature/retoots_in_feed_scorer.ts
@@ -11,12 +11,20 @@ export default class RetootsInFeedScorer extends FeatureScorer {
         super(WeightName.RETOOTED_IN_FEED);
     }
 
-    // TODO: should this subtract one for the retoot that put the toot in the user's feed?
     async _score(toot: Toot) {
         if (toot.reblog) {
             const reblog = toot.reblog;
             const nonAuthorRetoots = reblog.reblogsBy.filter((acct) => acct.webfingerURI != reblog.account.webfingerURI);
-            return Math.pow(nonAuthorRetoots.length, 2);
+            let retootCount = nonAuthorRetoots.length;
+            if (toot.reblog.isFollowed) retootCount += 1;  // add 1 if both reblog & toot are followed accounts
+            retootCount -= 1;  // Subtract 1 so that normal retoots aren't boosted unnecessarily
+
+            if (retootCount < 0) {
+                console.warn(`[${this.constructor.name}] Negative retoot count ${retootCount} for toot:`, toot);
+                return 0;
+            }
+
+            return Math.pow(retootCount, 2);
         } else {
             return 0;
         }