+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # NEXT RELEASE
 * RetootsInFeedScorer can use Toot.reblogsBy.length to score (and therefore doesn't need to be a FeedScorer)
-* Remove TrendingLinksScorer.trendingLinks property
+* Remove `TrendingLinksScorer.trendingLinks` property
+* Stop mutating the Toot objects in FollowedTagsScorer (now happens in Toot.setDependentProperties())
 
 ### v0.20.7
 * Consolidate all `Toot` post processing to `Toot.setDependentProperties()` method