+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # NEXT RELEASE
-* RetootsInFeedScorer can use Toot.reblogsBy.length to score (and therefore doesn't need to be a FeedScorer)
+* `RetootsInFeedScorer` can use Toot.reblogsBy.length to score (and therefore doesn't need to be a FeedScorer)
+* `RetootsInFeedScorer` scores with the square of the number of retooting accounts the user follows
 * Remove `TrendingLinksScorer.trendingLinks` property
 * Stop mutating the Toot objects in FollowedTagsScorer (now happens in Toot.setDependentProperties())