+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 * `RetootsInFeedScorer` can use Toot.reblogsBy.length to score (and therefore doesn't need to be a FeedScorer)
 * `RetootsInFeedScorer` scores with the square of the number of retooting accounts the user follows
 * Remove `TrendingLinksScorer.trendingLinks` property
-* Stop mutating the Toot objects in FollowedTagsScorer (now happens in Toot.setDependentProperties())
+* Stop mutating the `Toot` objects in `FollowedTagsScorer` (now happens in `Toot.setDependentProperties()`)
+* Set `Toot.muted` in Toot.setDependentProperties()
 
 ### v0.20.7
 * Consolidate all `Toot` post processing to `Toot.setDependentProperties()` method