+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # NEXT RELEASE
-* `RetootsInFeedScorer` can use Toot.reblogsBy.length to score (and therefore doesn't need to be a FeedScorer)
+* `RetootsInFeedScorer` can use `Toot.reblogsBy.length` to score (and therefore doesn't need to be a `FeedScorer`)
 * `RetootsInFeedScorer` scores with the square of the number of retooting accounts the user follows
 * Remove `TrendingLinksScorer.trendingLinks` property
 * Stop mutating the `Toot` objects in `FollowedTagsScorer` (now happens in `Toot.setDependentProperties()`)
-* Set `Toot.muted` in Toot.setDependentProperties()
+* Set `Toot.muted` where necessary in `Toot.setDependentProperties()`
 * Add Toot.setDependentProps() static method to set props for array of Toots. move the setting of those props to the calls like MastoApi.fetchHomeFeed() that do the initial Toot object building
 
 ### v0.20.7